backend/requirements.txt
# AI自动化工作流 - Python 依赖

# Web框架
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9

# 数据验证
pydantic>=2.6.0
pydantic-settings>=2.2.0
email-validator>=2.1.0

# 数据库
sqlalchemy>=2.0.0

# 认证
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# HTTP客户端
httpx>=0.27.0

# 定时任务
apscheduler>=3.10.0

# 环境变量
python-dotenv>=1.0.0
